Change bg img width and height from max to screen
If the background image has utility classes w-screen and h-screen,
then there is no extra grey space when the user's screen is
long vertical and short horizontal, but there is extra grey space
when the user's screen is long horizontal and short vertical.

The opposite is true for the utility classes w-max and h-max.